requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0
transformers>=4.30.0

# Data Processing
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
opencv-python>=4.8.0
torchmetrics>=1.0.0
huggingface_hub>=0.16.0